# select: do not shuffle channel ids in `Select::maybe_try_select`