I was browsing through the source code and I did not understand why, in the `Select::maybe_try_select` function, the order of the channel ids, in which they are `try`'d should be random. This patch removes this randomization, so you can see what (pieces of code) I mean.