@dsprenkels 2016-04-08
Because one of the benchmarks fails horribly on this patch, I am going to close this by now.