@BurntSushi 2016-04-08
Shuffling is actually pretty critical. And it's part of the public API contract. From the docs for `chan_select!`:

> It first examines all send and receive operations. If one or more of them can succeed without blocking, then it randomly selects one, executes the operation and runs the code in the corresponding arm.

The specific reason why this is done is to avoid starvation. For example, if `chan_select!` always chose the same channel at every synchronization point, then it could be that some channels *never* synchronize, which is bad.

AFAIK, this is a standard heuristic in CSP libraries. Repy's original Concurrent ML paper talks about it at least, I think. Go probably also does it.

I'm sure it's not currently as *efficient* as it could be though!