@dsprenkels 2016-04-08
That actually makes a lot of sense. I had originally expected that a channel would not synchronize when it gets empty, and so we would be able to look at the next one. This obviously does not happen, so starvation will indeed be the case here.

Then shuffling this list is a good step. If efficiency is important here, we could just select the first channel randomly, and continue through the list from then on.