@BurntSushi 2016-04-08
Well, even when that's not the case, if the same two channels are always ready for synchronization whenever select runs, then it would always pick the same channel without some kind of shuffling. (Which would lead to starvation.)