@BurntSushi 2016-04-08
@dsprenkels Do the benchmarks still look OK? Otherwise this looks good I think.