@dsprenkels 2016-04-08
(As expected), the benchmarks for all unbuffered benchmarks run somewhat slower, but most buffered benchmarks run a lot faster.

_However, something seems to go wrong in the buffered *single consumer* benchmarks. With this in mind, this patch is currently far from an improvement._

**Master**
```
test mpsc_chan_async                         ... bench:     760,109 ns/iter (+/- 4,103,188) = 1 MB/s
test mpsc_chan_sync_buffered                 ... bench:  15,767,711 ns/iter (+/- 3,964,848)
test mpsc_chan_sync_buffered_all             ... bench:     459,241 ns/iter (+/- 56,280) = 2 MB/s
test mpsc_chan_sync_unbuffered               ... bench:  24,971,355 ns/iter (+/- 4,384,348)
test select_no_init_chan_async               ... bench:   7,466,212 ns/iter (+/- 5,730,615)
test select_no_init_chan_sync_buffered       ... bench:  15,256,673 ns/iter (+/- 10,804,908)
test select_no_init_chan_sync_buffered_all   ... bench:   6,831,703 ns/iter (+/- 5,930,661)
test select_no_init_chan_sync_unbuffered     ... bench:  15,347,068 ns/iter (+/- 7,629,978)
test select_with_init_chan_async             ... bench:   7,002,994 ns/iter (+/- 6,348,005)
test select_with_init_chan_sync_buffered     ... bench:  17,974,850 ns/iter (+/- 8,736,832)
test select_with_init_chan_sync_buffered_all ... bench:   7,994,633 ns/iter (+/- 5,812,186)
test select_with_init_chan_sync_unbuffered   ... bench:  18,646,250 ns/iter (+/- 14,772,533)
test spsc_chan_async                         ... bench:   3,322,565 ns/iter (+/- 4,053,409)
test spsc_chan_sync_buffered                 ... bench:  16,355,497 ns/iter (+/- 21,629,817)
test spsc_chan_sync_buffered_all             ... bench:     404,213 ns/iter (+/- 3,932,395) = 2 MB/s
test spsc_chan_sync_unbuffered               ... bench:  11,632,866 ns/iter (+/- 12,255,797)
test std_mpsc_chan_async                     ... bench:     295,921 ns/iter (+/- 451,862) = 3 MB/s
test std_mpsc_chan_sync_buffered             ... bench:  23,610,042 ns/iter (+/- 20,420,668)
test std_mpsc_chan_sync_buffered_all         ... bench:     204,664 ns/iter (+/- 14,426) = 4 MB/s
test std_mpsc_chan_sync_unbuffered           ... bench:  22,179,166 ns/iter (+/- 12,572,612)
test std_spsc_chan_async      ==!!!!!!!!==   ... bench:    *196,200*ns/iter (+/- 109,231) = 5 MB/s
test std_spsc_chan_sync_buffered             ... bench:  23,995,305 ns/iter (+/- 22,014,797)
test std_spsc_chan_sync_buffered_all         ... bench:     213,297 ns/iter (+/- 648,828) = 4 MB/s
test std_spsc_chan_sync_unbuffered           ... bench:   8,835,259 ns/iter (+/- 20,673,593)
```

**After patch**
```
test mpsc_chan_async                         ... bench:     469,554 ns/iter (+/- 78,060) = 2 MB/s
test mpsc_chan_sync_buffered                 ... bench:  15,515,844 ns/iter (+/- 7,422,272)
test mpsc_chan_sync_buffered_all             ... bench:     460,473 ns/iter (+/- 105,530) = 2 MB/s
test mpsc_chan_sync_unbuffered               ... bench:  26,473,704 ns/iter (+/- 3,866,253)
test select_no_init_chan_async               ... bench:   5,391,971 ns/iter (+/- 6,512,831)
test select_no_init_chan_sync_buffered       ... bench:  15,661,369 ns/iter (+/- 15,924,706)
test select_no_init_chan_sync_buffered_all   ... bench:   3,229,798 ns/iter (+/- 4,703,352)
test select_no_init_chan_sync_unbuffered     ... bench:  14,460,452 ns/iter (+/- 10,480,913)
test select_with_init_chan_async             ... bench:   4,886,006 ns/iter (+/- 5,192,072)
test select_with_init_chan_sync_buffered     ... bench:  14,274,815 ns/iter (+/- 8,843,767)
test select_with_init_chan_sync_buffered_all ... bench:   2,412,597 ns/iter (+/- 7,064,717)
test select_with_init_chan_sync_unbuffered   ... bench:  15,563,561 ns/iter (+/- 13,823,595)
test spsc_chan_async                         ... bench:   2,555,373 ns/iter (+/- 4,070,408)
test spsc_chan_sync_buffered                 ... bench:  12,974,285 ns/iter (+/- 13,482,108)
test spsc_chan_sync_buffered_all             ... bench:   1,957,220 ns/iter (+/- 5,784,124)
test spsc_chan_sync_unbuffered               ... bench:  19,701,297 ns/iter (+/- 8,574,394)
test std_mpsc_chan_async                     ... bench:   2,923,443 ns/iter (+/- 2,827,214)
test std_mpsc_chan_sync_buffered             ... bench:  12,118,643 ns/iter (+/- 18,853,889)
test std_mpsc_chan_sync_buffered_all         ... bench:   2,908,142 ns/iter (+/- 1,977,568)
test std_mpsc_chan_sync_unbuffered           ... bench:  14,205,326 ns/iter (+/- 9,913,628)
test std_spsc_chan_async      ==!!!!!!!!==   ... bench:  *3,867,831*ns/iter (+/- 5,189,367)
test std_spsc_chan_sync_buffered             ... bench:  16,157,883 ns/iter (+/- 20,936,732)
test std_spsc_chan_sync_buffered_all         ... bench:     271,048 ns/iter (+/- 773,941) = 3 MB/s
test std_spsc_chan_sync_unbuffered           ... bench:  12,560,218 ns/iter (+/- 20,416,679)
```