@dsprenkels 2016-04-08
@BurntSushi The changes in speed are so minimal in the cost of code readability, that I think this patch will not be an improvement, so I am closing it (once again).