@BurntSushi 2016-04-08
@dsprenkels I think I probably agree. To be honest, I don't think this library is really at a point where this level of optimization makes that much sense. I suspect it needs more careful tuning elsewhere (probably in how much synchronization is actually happening) before something like this will matter. I know that I personally focused almost all of my attention on getting it correct, so I bet there's a lot of perf being left on the table as far as synchronization is concerned. For example, there's probably a lot of thrashing going on right now between selects and channels, since they each wake each other up quite aggressively!